+++ tests/test_axis.py
@@ -18,6 +18,28 @@ class TestChart(XMLTest):
             (10, 4)
         ]
 
+    def test_ticks(self):
+        chart = leather.Chart()
+        chart.add_dots(self.data)
+
+        axis = leather.Axis(ticks=6)
+        chart.set_x_axis(axis)
+
+        svg = self.render_chart(chart)
+
+        self.assertTickLabels(svg, 'bottom', ['2.0', '4.0', '6.0', '8.0', '10.0', '0.0'])
+
+    def test_tick_values(self):
+        chart = leather.Chart()
+        chart.add_dots(self.data)
+
+        axis = leather.Axis(tick_values=[-12, 0, 17, 44, 87, 99])
+        chart.set_x_axis(axis)
+
+        svg = self.render_chart(chart)
+
+        self.assertTickLabels(svg, 'bottom', ['-12', '17', '44', '87', '99', '0'])
+
     def test_tick_formatter(self):
         chart = leather.Chart()
         chart.add_dots(self.data)